# src/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Scenor API",
    "description": "Auto-generated API docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication endpoints"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register user",
        "description": "",
        "responses": {
          "201": {
            "description": "User created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "",
        "responses": {
          "200": {
            "description": "Authenticated user with token"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current user",
        "description": "",
        "responses": {
          "200": {
            "description": "Current user"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Update current user",
        "description": "",
        "responses": {
          "200": {
            "description": "Updated user"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegisterUserRequest": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "username": {
                "type": "string",
                "example": "w00xy"
              },
              "password": {
                "type": "string",
                "example": "strong-password"
              }
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "password": {
                "type": "string",
                "example": "strong-password"
              }
            }
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "new@example.com"
              },
              "username": {
                "type": "string",
                "example": "new-name"
              },
              "password": {
                "type": "string",
                "example": "new-password"
              }
            }
          }
        }
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "example": "w00xy"
          },
          "token": {
            "type": "string",
            "example": "jwt-token"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use `Bearer <token>`"
      }
    }
  }
}